+++ R/futurize.R
@@ -11,13 +11,15 @@
 #' @param options,\ldots Named options, passed to [futurize_options()],
 #' controlling how futures are resolved.
 #' 
+#' @param when If TRUE (default), the expression is futurized, otherwise not.
+#'
+#' @param eval If TRUE (default), the futurized expression is evaluated,
+#' other it is returned.
+#'
 #' @param flavor Flavor of futurize transpiler to use.
 #' If `"add-on"`, then registered transpilers for packages such as
 #' \pkg{future.apply} and \pkg{furrr} are used.
 #' If `"built-in"`, then built-in transpilers are used.
-#'
-#' @param eval If TRUE (default), the futurized expression is evaluated,
-#' other it is returned.
 #'
 #' @returns
 #' Returns the value of the evaluated expression `expr`.
@@ -25,6 +27,23 @@
 #' If `expr` is a TRUE or FALSE, then a logical is returned indicating 
 #' whether futurization was previously enabled or disabled.
 #'
+#'
+#' @section Conditionally futurization:
+#' It is possible to control whether futurization should take place at
+#' run-time. For example,
+#'
+#' ```r
+#' y <- lapply(xs, fun) |> futurize(when = { length(xs) >= 10 })
+#' ```
+#'
+#' will be futurized, unless `length(xs)` less than ten, in case it is
+#' evaluated as:
+#'
+#' ```r
+#' y <- lapply(xs, fun)
+#' ```
+#'
+#' 
 #' @section Disable and re-enable all futurization:
 #' It is possible to globally disable the effect of all `futurize()` calls
 #' by calling `futurize(FALSE)`. The effect is as if `futurize()` was never
@@ -52,7 +71,7 @@
 #' @aliases parallelize
 #' @importFrom future future value
 #' @export
-futurize <- function(expr, substitute = TRUE, options = futurize_options(...), ..., flavor = c("add-on", "built-in"), eval = TRUE, envir = parent.frame()) {
+futurize <- function(expr, substitute = TRUE, options = futurize_options(...), ..., when = TRUE, eval = TRUE, flavor = c("add-on", "built-in"), envir = parent.frame()) {
   if (substitute) expr <- substitute(expr)
   debug <- isTRUE(getOption("futurize.debug"))
   if (debug) {
@@ -67,8 +86,12 @@ futurize <- function(expr, substitute = TRUE, options = futurize_options(...), .
     return(invisible(.futurize))
   }
 
-  ## Do nothing?
-  if (!.futurize) {
+  stopifnot(
+    is.logical(when), length(when) == 1L, !is.na(when)
+  )
+
+  ## Futurize?
+  if (!when || !.futurize) {
     if (eval) {
       if (debug) mdebug("Evaluate call expression")
       return(eval(expr, envir = envir))
@@ -78,10 +101,6 @@ futurize <- function(expr, substitute = TRUE, options = futurize_options(...), .
     }
   }
   
-  stopifnot(
-    is.language(expr),
-    is.call(expr)
-  )
 
   flavor <- match.arg(flavor, several.ok = FALSE)